+++ packages/chrome-extension/src/internal/clerk.ts
@@ -1,14 +1,15 @@
-import { Clerk } from '@clerk/clerk-js';
+import { Clerk } from '@clerk/clerk-js/no-rhc';
 import { DEV_BROWSER_JWT_KEY } from '@clerk/shared/devBrowser';
 import { parsePublishableKey } from '@clerk/shared/keys';
 import browser from 'webextension-polyfill';
 
 import { SCOPE, type Scope } from '../types';
-import { AUTH_HEADER, CLIENT_JWT_KEY, DEFAULT_LOCAL_HOST_PERMISSION } from './constants';
-import type { GetClientCookieParams } from './utils/cookies';
+import { CLIENT_JWT_KEY, DEFAULT_LOCAL_HOST_PERMISSION } from './constants';
 import { assertPublishableKey } from './utils/errors';
 import { JWTHandler } from './utils/jwt-handler';
-import { getValidPossibleManifestHosts, validateHostPermissionExistence, validateManifest } from './utils/manifest';
+import { validateManifest } from './utils/manifest';
+import { requestHandler } from './utils/request-handler';
+import { responseHandler } from './utils/response-handler';
 import { BrowserStorageCache, type StorageCache } from './utils/storage';
 
 export let clerk: Clerk;
@@ -22,18 +23,26 @@ export type CreateClerkClientOptions = {
   publishableKey: string;
   scope?: Scope;
   storageCache?: StorageCache;
-  syncSessionWithTab?: boolean;
+  syncHost?: string;
 };
 
 export async function createClerkClient({
   publishableKey,
   scope,
   storageCache = BrowserStorageCache,
-  syncSessionWithTab = false,
+  syncHost,
 }: CreateClerkClientOptions): Promise<Clerk> {
-  if (clerk) {
+  if (scope === SCOPE.BACKGROUND) {
+    Clerk.mountComponentRenderer = undefined;
+  }
+
+  // Don't cache background scripts as it can result in out-of-sync client information.
+  if (clerk && scope !== SCOPE.BACKGROUND) {
     return clerk;
   }
+
+  // Sync is enabled if a `syncHost` is provided
+  const sync = Boolean(syncHost);
 
   // Parse publishableKey and assert it's present/valid, throw if not
   const key = parsePublishableKey(publishableKey);
@@ -44,79 +53,26 @@ export async function createClerkClient({
 
   // Will throw if manifest is invalid
   validateManifest(manifest, {
-    sync: syncSessionWithTab,
-    background: scope === SCOPE.background,
+    background: scope === SCOPE.BACKGROUND,
+    sync,
   });
 
-  let jwt: JWTHandler | undefined;
+  // Set up JWT handler and attempt to get JWT from storage on initialization
+  const url = syncHost ? syncHost : DEFAULT_LOCAL_HOST_PERMISSION;
 
-  if (syncSessionWithTab) {
-    const hostHint = isProd ? key.frontendApi : DEFAULT_LOCAL_HOST_PERMISSION;
-    const validHosts = getValidPossibleManifestHosts(manifest);
+  const jwtOptions = {
+    frontendApi: key.frontendApi,
+    name: isProd ? CLIENT_JWT_KEY : DEV_BROWSER_JWT_KEY,
+    sync,
+    url,
+  };
 
-    // Will throw if manifest host_permissions doesn't contain a valid host
-    validateHostPermissionExistence(validHosts, hostHint);
-
-    // Set up cookie params based on environment
-    const getClientCookieParams: GetClientCookieParams = isProd
-      ? {
-          urls: `https://${key.frontendApi}`,
-          name: CLIENT_JWT_KEY,
-        }
-      : {
-          urls: validHosts,
-          name: DEV_BROWSER_JWT_KEY,
-        };
-
-    // Set up JWT handler and attempt to get JWT from storage on initialization
-    jwt = JWTHandler(storageCache, { ...getClientCookieParams, frontendApi: key.frontendApi, sync: true });
-  } else {
-    jwt = JWTHandler(storageCache, { frontendApi: key.frontendApi, sync: false });
-  }
-
-  if (!jwt) {
-    throw new Error('StorageCache could not be initialized.'); // TODO: Update error
-  }
+  const jwt = JWTHandler(storageCache, jwtOptions);
 
   // Create Clerk instance
   clerk = new Clerk(publishableKey);
-
-  // Append appropriate query params to all Clerk requests
-  clerk.__unstable__onBeforeRequest(async requestInit => {
-    requestInit.credentials = 'omit';
-
-    const currentJWT = await jwt.get();
-
-    if (!currentJWT) {
-      requestInit.url?.searchParams.append('_is_native', '1');
-      return;
-    }
-
-    if (isProd) {
-      requestInit.url?.searchParams.append('_is_native', '1');
-      (requestInit.headers as Headers).set('authorization', `Bearer ${currentJWT}`);
-    } else {
-      requestInit.url?.searchParams.append('__clerk_db_jwt', currentJWT);
-    }
-  });
-
-  // Store updated JWT in StorageCache on Clerk responses
-  clerk.__unstable__onAfterResponse(async (_, response) => {
-    const authHeaderkey = isProd ? AUTH_HEADER.production : AUTH_HEADER.development;
-    const authHeader = response?.headers.get(authHeaderkey);
-
-    if (authHeader?.startsWith('Bearer')) {
-      const newJWT = authHeader.split(' ')[1] || undefined;
-
-      if (newJWT) {
-        await jwt.set(newJWT);
-      } else {
-        await jwt.remove();
-      }
-    } else if (authHeader) {
-      await jwt.set(authHeader);
-    }
-  });
+  clerk.__unstable__onAfterResponse(responseHandler(jwt, { isProd }));
+  clerk.__unstable__onBeforeRequest(requestHandler(jwt, { isProd }));
 
   return clerk;
 }