+++ packages/clerk-js/rspack.config.js
@@ -11,6 +11,7 @@ const isDevelopment = mode => !isProduction(mode);
 
 const variants = {
   clerk: 'clerk',
+  clerkNoRHC: 'clerk.no-rhc', // Omit Remotely Hosted Code
   clerkBrowser: 'clerk.browser',
   clerkHeadless: 'clerk.headless',
   clerkHeadlessBrowser: 'clerk.headless.browser',
@@ -18,6 +19,7 @@ const variants = {
 
 const variantToSourceFile = {
   [variants.clerk]: './src/index.ts',
+  [variants.clerkNoRHC]: './src/index.ts',
   [variants.clerkBrowser]: './src/index.browser.ts',
   [variants.clerkHeadless]: './src/index.headless.ts',
   [variants.clerkHeadlessBrowser]: './src/index.headless.browser.ts',
@@ -27,9 +29,10 @@ const variantToSourceFile = {
  *
  * @param {object} config
  * @param {'development'|'production'} config.mode
+ * @param {boolean} [config.disableRHC=false]
  * @returns { import('@rspack/cli').Configuration }
  */
-const common = ({ mode }) => {
+const common = ({ mode, disableRHC = false }) => {
   return {
     mode,
     resolve: {
@@ -39,7 +42,9 @@ const common = ({ mode }) => {
     },
     plugins: [
       new rspack.DefinePlugin({
+        __BUILD_DISABLE_RHC__: JSON.stringify(disableRHC),
         __DEV__: isDevelopment(mode),
+        __BUILD_FLAG_ACCOUNTLESS_UI__: isDevelopment(mode),
         __PKG_VERSION__: JSON.stringify(packageJSON.version),
         __PKG_NAME__: JSON.stringify(packageJSON.name),
         BUILD_ENABLE_NEW_COMPONENTS: JSON.stringify(process.env.BUILD_ENABLE_NEW_COMPONENTS),
@@ -400,12 +405,63 @@ const prodConfig = ({ mode, env, analysis }) => {
     },
   });
 
+  const clerkEsmNoRHC = merge(
+    entryForVariant(variants.clerkNoRHC),
+    common({ mode, disableRHC: true }),
+    commonForProd(),
+    commonForProdBundled(),
+    {
+      experiments: {
+        outputModule: true,
+      },
+      output: {
+        filename: '[name].mjs',
+        libraryTarget: 'module',
+      },
+      plugins: [
+        // Include the lazy chunks in the bundle as well
+        // so that the final bundle can be imported and bundled again
+        // by a different bundler, eg the webpack instance used by react-scripts
+        new rspack.optimize.LimitChunkCountPlugin({
+          maxChunks: 1,
+        }),
+      ],
+      optimization: {
+        splitChunks: false,
+      },
+    },
+  );
+
+  const clerkCjsNoRHC = merge(
+    entryForVariant(variants.clerkNoRHC),
+    common({ mode, disableRHC: true }),
+    commonForProd(),
+    commonForProdBundled(),
+    {
+      output: {
+        filename: '[name].js',
+        libraryTarget: 'commonjs',
+      },
+      plugins: [
+        // Include the lazy chunks in the bundle as well
+        // so that the final bundle can be imported and bundled again
+        // by a different bundler, eg the webpack instance used by react-scripts
+        new rspack.optimize.LimitChunkCountPlugin({
+          maxChunks: 1,
+        }),
+      ],
+      optimization: {
+        splitChunks: false,
+      },
+    },
+  );
+
   // webpack-bundle-analyzer only supports a single build, use clerkBrowser as that's the default build we serve
   if (analysis) {
     return [clerkBrowser];
   }
 
-  return [clerkBrowser, clerkHeadless, clerkHeadlessBrowser, clerkEsm, clerkCjs];
+  return [clerkBrowser, clerkHeadless, clerkHeadlessBrowser, clerkEsm, clerkEsmNoRHC, clerkCjs, clerkCjsNoRHC];
 };
 
 /**
@@ -476,6 +532,12 @@ const devConfig = ({ mode, env }) => {
       entryForVariant(variants.clerkBrowser),
       isSandbox ? { entry: { sandbox: './sandbox/app.js' } } : {},
       common({ mode }),
+      commonForDev(),
+    ),
+    // prettier-ignore
+    [variants.clerkBrowserNoRHC]: merge(
+      entryForVariant(variants.clerkBrowserNoRHC),
+      common({ mode, disableRHC: true }),
       commonForDev(),
     ),
     [variants.clerkHeadless]: merge(