+++ packages/clerk-js/src/ui/Components.tsx
@@ -1,7 +1,7 @@
 import { useSafeLayoutEffect } from '@clerk/shared/react';
 import { createDeferredPromise } from '@clerk/shared/utils';
 import type {
-  __experimental_UserVerificationProps,
+  __internal_UserVerificationProps,
   Appearance,
   Clerk,
   ClerkOptions,
@@ -79,7 +79,7 @@ export type ComponentControls = {
       : T extends 'signUp'
         ? SignUpProps
         : T extends 'userVerification'
-          ? __experimental_UserVerificationProps
+          ? __internal_UserVerificationProps
           : T extends 'waitlist'
             ? WaitlistProps
             : UserProfileProps,
@@ -125,7 +125,7 @@ interface ComponentsState {
   signInModal: null | SignInProps;
   signUpModal: null | SignUpProps;
   userProfileModal: null | UserProfileProps;
-  userVerificationModal: null | __experimental_UserVerificationProps;
+  userVerificationModal: null | __internal_UserVerificationProps;
   organizationProfileModal: null | OrganizationProfileProps;
   createOrganizationModal: null | CreateOrganizationProps;
   blankCaptchaModal: null;
@@ -517,11 +517,13 @@ const Components = (props: ComponentsProps) => {
           </LazyImpersonationFabProvider>
         )}
 
-        {state.options?.__internal_claimAccountlessKeysUrl && (
-          <LazyImpersonationFabProvider globalAppearance={state.appearance}>
-            <AccountlessPrompt url={state.options.__internal_claimAccountlessKeysUrl} />
-          </LazyImpersonationFabProvider>
-        )}
+        {__BUILD_FLAG_ACCOUNTLESS_UI__
+          ? state.options?.__internal_claimAccountlessKeysUrl && (
+              <LazyImpersonationFabProvider globalAppearance={state.appearance}>
+                <AccountlessPrompt url={state.options.__internal_claimAccountlessKeysUrl} />
+              </LazyImpersonationFabProvider>
+            )
+          : null}
 
         <Suspense>{state.organizationSwitcherPrefetch && <OrganizationSwitcherPrefetch />}</Suspense>
       </LazyProviders>